Replace PostCard component with tsx
