--- src/components/Modules/PostCard/index.ts
@@ -1,3 +1,0 @@
-import PostCard from './PostCard.vue'
-
-export default PostCard
